fix: Replaced LLTurnIn with TurnIn
Replaced the LLTurnIn tag with the stock TurnIn tag. 7.1 introduced a new window into the turn in process which was causing the tag to get stuck when attempting to turn in quests. As the stock TurnIn tag has been upgraded to meet our needs, there's no functional reason to continue using LLTurnIn.